Move config out of support into `src`
